requirements.txt
flask
feedparser
loguru
requests
openai
hydra-core